ui: split sidebar into own component
